requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
joblib==1.3.2
matplotlib==3.8.0
seaborn==0.13.0
plotly==5.17.0
altair==5.1.2
opencv-python==4.8.1.78
Pillow==10.1.0
requests==2.31.0

# Audio processing
librosa==0.10.1
soundfile==0.12.1
sounddevice==0.4.6

# Deep learning and computer vision
torch==2.1.0
torchvision==0.16.0
face-alignment==1.3.5

# Additional utilities
websockets==12.0
paho-mqtt==1.6.1
tqdm==4.66.1
scipy==1.11.3
pyyaml==6.0.1
python-multipart==0.0.6
protobuf==4.25.1

# For audio format conversion (needs system ffmpeg)
# System dependency: ffmpeg
supabase==1.4.0
python-dotenv==1.0.0